backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
PyPDF2==3.0.1
pytesseract==0.3.13
Pillow==11.1.0
pymongo==4.10.1
spacy==3.8.2
requests==2.32.3
beautifulsoup4==4.12.3
